refactor: BottomSheetProvider 파일에 context 통합 작성 #474
